+++ decidim-core/config/routes.rb
@@ -124,5 +124,5 @@ Decidim::Core::Engine.routes.draw do
     get "/me" => "doorkeeper/credentials#me"
   end
 
-  root to: "pages#show", id: "home"
+  root to: "homepage#show"
 end